pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tour"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
